+++ alphanumeric.js
@@ -1,5 +1,0 @@
-const numbers = require('./numbers');
-const lowercase = require('./lowercase');
-const uppercase = require('./uppercase');
-
-module.exports = numbers + lowercase + uppercase;