+++ numbers.js
@@ -1,1 +1,0 @@
-module.exports = '0123456789';