+++ src/nolookalikes.js
@@ -1,0 +1,1 @@
+export const nolookalikes = '346789ABCDEFGHJKLMNPQRTUVWXYabcdefghijkmnpqrtwxyz';