+++ src/numbers.js
@@ -1,0 +1,1 @@
+export const numbers = '0123456789';